updated versionof mlserver

+++ pipelines/mlserver-final/audio-qa/seldon-core-version/nodes/audio/load-test-grpc.py
@@ -35,7 +35,7 @@ ds = load_dataset(
     "hf-internal-testing/librispeech_asr_demo",
     "clean",
     split="validation")
-data = ds[0]["audio"]["array"].astype(np.float32)
+data = ds[0]["audio"]["array"]
 data_shape = [len(data)]
 custom_parameters = {'custom_1': 'test_1'}
 data_1 = Data(
@@ -49,7 +49,7 @@ ds = load_dataset(
     "hf-internal-testing/librispeech_asr_demo",
     "clean",
     split="validation")
-data = ds[0]["audio"]["array"].astype(np.float32)
+data = ds[0]["audio"]["array"]
 data_shape = [len(data)]
 custom_parameters = {'custom_2': 'test_2'}
 data_2 = Data(
